requirement.txt
streamlit==1.38.0
openai==1.45.0
openai-whisper==20240930
sounddevice==0.5.0
wavio==0.0.8
streamlit-drawable-canvas==0.9.3
numpy==1.26.4
requests==2.32.3
gtts==2.5.3
pygame==2.5.2  # Downgraded to avoid ImpImporter issue